chatbot-server/requirements.txt
fastapi
uvicorn
google-auth
google-genai
google-cloud-aiplatform
pydantic 